+++ src/lib/helper/helper_system.cc
@@ -4,7 +4,7 @@
 //    (See accompanying file LICENSE or copy at
 //          http://www.boost.org/LICENSE_1_0.txt)
 
-#include <string>
+#include <cstring>
 #include <clocale>
 #include <iostream>
 #include <cstdlib>